+++ R/get_series_table.R
@@ -13,7 +13,7 @@
 #' @param parse_values optional boolean. If set to `true` (default) it will
 #' attempt to parse the contents of `value` and cast numeric strings as numeric
 #' values. If set to `false` it will keep return a `value` column of strings.
-#' @param ... additional parameters to pass to [`bls_request`]
+#' @param ... additional parameters to pass to [`get_series`]
 #'
 #' @return a tibble of observations or `NA` if the request had zero results.
 #'
@@ -31,32 +31,12 @@ get_series_table <- function(
   series_id, api_key,
   start_year=NA, end_year=NA, year_limit=20, parse_values=TRUE, ...
   ){
+
   if(is.na(api_key)) year_limit <- 10
-  #auto-magically paginate results
-  if(!is.na(start_year) && (end_year - start_year) >= year_limit ){
-    rlang::inform( sprintf(
-      'Year %i to %i is longer than %i API limit. Performing multiple requests',
-      start_year, end_year, year_limit
-    ) )
 
-    tables <- lapply(
-      seq(0, ceiling(((end_year - start_year) + 1) / year_limit) - 1) * year_limit,
-      function(x) get_series_table(
-        series_id = series_id,
-        api_key = api_key,
-        start_year = start_year+x,
-        end_year = min(end_year, start_year+(x+(year_limit-1))),
-        parse_values = parse_values,
-        ...
-      )
-    )
-
-    return(purrr::reduce(purrr::keep(tables, is.data.frame), dplyr::union_all))
-  }
-
-  series <- get_series(series_id, start_year, end_year, api_key=api_key, ...)
+  series <- get_series(series_id, start_year, end_year, year_limit, api_key=api_key, ...)
   if(length(series$data) > 0)
-    return(data_as_table(series$data,parse_values))
+    return(data_as_table(series$data, parse_values))
 
   NA
 }
@@ -76,7 +56,7 @@ get_series_table <- function(
 #' @param parse_values optional boolean. If set to `true` (default) it will
 #' attempt to parse the contents of `value` and cast numeric strings as numeric
 #' values. If set to `false` it will keep return a `value` column of strings.
-#' @param ... additional parameters to pass to [`query_n_series`]
+#' @param ... additional parameters to pass to [`get_n_series`]
 #'
 #' @return a list of tibbles. Series requests which return observations will be
 #' a tibble. Series with no observations will be `NA`
@@ -103,34 +83,9 @@ get_series_tables <- function(
   series_ids, api_key,
   start_year=NA, end_year=NA, year_limit=20, parse_values=TRUE, ...
 ){
-  if(is.na(api_key)) year_limit <- 10
-  #auto-magically paginate results
-  if(!is.na(start_year) && (end_year - start_year) >= year_limit ){
-    rlang::inform( sprintf(
-      'Year %i to %i is longer than %i API limit. Performing multiple requests',
-      start_year, end_year, year_limit
-    ) )
 
-    pages <- lapply(
-      seq(0, ceiling(((end_year - start_year) + 1) / year_limit) - 1) * year_limit,
-      function(x) get_series_tables(
-        series_ids = series_ids,
-        api_key = api_key,
-        start_year = start_year+x,
-        end_year = min(end_year, start_year+(x+(year_limit-1))),
-        parse_values = parse_values,
-        ...
-      )
-    )
-    #invert the indices so the map reduce is simpler and keep only
-    tables <- purrr::map(purrr::transpose(pages), purrr::discard, rlang::is_na)
-
-    return( purrr::map(tables, ~purrr::reduce(rev(.x), dplyr::union_all) ) )
-  }
-
-  series_data <- purrr::map(
-    get_n_series(series_ids, api_key, start_year, end_year, ...), 'data'
-  )
+  series <- get_n_series(series_ids, api_key, start_year, end_year, year_limit, ...)
+  series_data <- purrr::map(series, 'data')
   purrr::modify_if(
     purrr::modify_if(series_data, rlang::is_empty, ~NA),
     purrr::negate(rlang::is_na),