+++ tests/testthat/test-requests.R
@@ -5,6 +5,36 @@ check_test_key <- function(){
   skip_if(identical(test_key, ""), "'BLS_API_TEST_KEY' not set.")
 }
 
+test_that('validate_years works',{
+
+  expect_silent(.validate_years(NA, NA))
+  expect_error(.validate_years(2014, NA), 'both')
+  expect_error(.validate_years(NA, 2014), 'both')
+  expect_error(.validate_years('2014','2015'), 'numeric')
+  expect_error(.validate_years(c(2014,2015),c(2016,2017)), 'scalar')
+  expect_error(.validate_years(2016,2015), 'greater')
+  expect_silent(.validate_years(2015,2016))
+  expect_silent(.validate_years(2015L,2016L))
+})
+
+ test_that('series_id_names works',{
+
+   expect_error(.series_id_names(c(1:5)), 'character vector')
+   expect_error(.series_id_names(c('foo')[-1]), 'empty')
+   expect_error(.series_id_names(c('a','b','')), 'empty')
+   expect_equal(.series_id_names(c('a','b','c')), c('a','b','c'))
+
+   expect_error(.series_id_names(list(1:3)), 'character vector')
+   expect_error(.series_id_names(list()), 'empty')
+   expect_error(.series_id_names(list(a='a', b='b',c='')), 'empty')
+   expect_equal(.series_id_names(list(a='d', b='e',c='f')), c('a','b','c'))
+#
+#   expect_error(.series_id_names(list(a='d', b='e','f')), 'named')
+#   expect_error(.series_id_names(list(a='d', a='e')), 'unique')
+#   expect_error(.series_id_names(list(a='d', a='e', ''='f')), 'empty')
+#
+ })
+
 test_that('live get_series_table request tests', {
   check_test_key()
   api_key <- Sys.getenv('BLS_API_TEST_KEY')