+++ NEWS.md
@@ -1,3 +1,38 @@
+# blsR 0.4.0
+
+  * Previously, the `get_n_series` documentation state that `...` was passed to
+  `query_n_series`. This was both inaccurate and inconsistent with the behavior 
+  of the other `get_*` functions. `get_n_series` arguments were updated to 
+  capture all of the `query_n_series` arguments and pass `...` to `bls_request`
+  * Previously, calling `get_n_series` for only 1 series would fail due to a
+  json encoding issue stemming from a length 1 vector being coerced to a scalar
+  within `httr::POST` in `bls_request`.
+  * Previously, `data_as_table` would return a variable number of columns based
+  on the shape of the input. This led to unpredictable output and caused issues
+  when merging multiple requests for different time spans. Going forward
+  `data_as_table` will return a tibble with only 4 columns (`year`, `period`, 
+  `periodName`, and `value`). Users that wish to access aspects, footnotes,
+  calculations, or the 'latest' flag should use `get_n_series`.
+  * As a result of the changes to `data_as_table`, `get_series_table` and 
+  `get_n_series_table` no longer include the `latest` or `footnotes` columns.
+  This fixes a condition where requesting aspects for a series without aspects
+  would cause the table to have zero rows and a condition where the `latest`
+  column could disappear between two identical calls if the data was updated.
+  * Previously, pagination for periods exceeding the API limits was done 
+  automatically in both `get_series_table` and `get_series_tables`. Pagination 
+  behavior has been renamed to 'spanning', removed from the previous points of
+  implementation and implemented as three spanning functions following a 
+  map-reduce pattern:
+    * `span_series_request` - encapsulates all of the spanning behavior
+    * `span_request_queries` - creates multiple conforming queries
+    * `reduce_spanned_responses` - merges the results of multiple queries
+  The spanning behavior is now implemented at `get_series` and `get_n_series`,
+  which removes the burden of merging multiple requests from the higher level
+  `get_series_table` and `get_n_series_table` functions and allows users to
+  create their own high-level functions to alter the presentation without having
+  to be aware of the multiple requests. Spanning behavior may be turned off by 
+  setting the `span` argument to `FALSE`.
+
 # blsR 0.3.2
 
   * Previously, if `get_series_table` or `get_series_tables` were called for a